Move handleFormSubmission outside
